requirements.txt
streamlit==1.45.0
openai==0.27.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.5
scikit-learn==1.4.2
